pyproject.toml
[tool.poetry]
name = "trab-proj-softw-1-backend"
version = "0.1.0"
description = ""
authors = ["Ivan Mendes <ivanmendes2005@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.2"
djangorestframework = "^3.17.1"
psycopg2-binary = "^2.9.11"
pillow = "^12.1.1"
django-environ = "^0.13.0"
gunicorn = "^25.2.0"
whitenoise = "^6.12.0"
django-storages = {extras = ["s3"], version = "^1.14.6"}
boto3 = "^1.42.76"
django-unfold = "^0.87.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
